leanEngine_app/requirements.txt
Flask>=0.10.1
leancloud-sdk>=1.0.9
#cython
#numpy
scikit-learn
blinker
logentries
bugsnag
hmmlearn
gevent
Theano
